backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fido2==1.1.2

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Cryptography
cryptography==41.0.7
pycryptodome==3.20.0
py-ecc==6.0.0
gmpy2==2.1.5

# Hyperledger Fabric SDK
hfc==1.0.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Logging and monitoring
structlog==24.1.0
prometheus-client==0.19.0
